Multi-window support
This patch adds the ability to open (and close) browser windows. The
application starts with a single browser window open. Additional windows
can be opened using the menu action or keyboard shortcut Ctrl+N. When
all windows are closed the browser will automatically exit. When all of
the tabs in a window are closed that window will automatically close,
similar to how other tabbed browsers behave.